experimentation: builds and runs but cspace invocations fail
The system compiles and runs (after commenting out some) incomplete
CSpace stuff, but there is something going wrong with the cspace
management stuff. Not quite clear what the problem is, because it seems
to be pretty much the same as the C equivalent. Will need to compare the
exact invocation used by C vs my system to figure out what's going on.