Initial implementation of BFS and loader (#7)
Finished an initial implementation of the boot file server and the smos-loader application. This involved the implementation of several new invocations and the implementation of notification forwarding between the root server and published servers using a ring buffer. There are also a number of QoL improvements that were done along the way.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>