Implemented the passing of environment variables (#12)
The loader runs an executable with the stack at a different location
than an application that was directly started by the root server (such
as the loader itself). This causes problems in the runtime crate, as
the stack top was assumed to be at a particular address. To address
this, envp was added in addition to argc and argv, and this is passed
into the application on the stack. This reduces the amount of hardcoded
data in the runtime crate, as other relevant memory locations, such as
the address of the IPC buffer and the shared buffer between the app and
the RS, are also passed in envp.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>