Implemented multi-level book-keeping for objects and views
Objects and views were linear, static arrays, which was problematic
as images got bigger, as the array would constantly need to be made
larger. This commit changes this to a more dynamic approach that looks
more like a per-object shadow page table. It is a bit rough around the
edges and probably uses more memory than it should at the moment.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>